trabajo final/tesis aon/tesis/software/requirements.txt
APScheduler==3.7.0
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.21.0
Pillow==8.2.0
pyparsing==2.4.7
pyserial==3.5
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
tzlocal==2.1
